api/requirements.txt
aioredis==1.3.1
anyio==3.5.0
APScheduler==3.9.1
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
ConfigArgParse==1.5.3
cronitor==4.6.0
cryptography==3.2
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.0
docopt==0.6.2
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
fastapi==0.74.1
filelock==3.7.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.33.3
gevent==21.12.0
geventhttpclient==2.0
greenlet==1.1.3
h11==0.13.0
hiredis==2.0.0
httptools==0.2.0
humanize==4.3.0
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.13.0
ipython==8.3.0
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.4.2
locust==2.11.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mlxtend==0.20.0
msgpack==1.0.4
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.2
orjson==3.6.7
packaging==21.3
pandas==1.4.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
PyJWT==2.4.0
pymongo==4.0.1
pyOpenSSL==20.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==22.3.0
redis==4.3.4
regex==2022.4.24
requests==2.27.1
roundrobin==0.0.4
scikit-learn==1.0.2
scipy==1.8.0
sentry-sdk==1.9.5
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.17.1
stomp.py==8.0.0
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.63.0
traitlets==5.2.0
typing-extensions==4.1.1
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.12
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.2
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
